Navbar component creation
redirected links to the proper path, imported a few assets from the assets folder. Also, set our initial states, interpolated styles from the styles folder with classname properties from tailwind. set the state in the onclick arrow function to empty and to scroll to the top of the page (0,0). imported the logo and changed the title, logo the title bar. made a few changes to the navigation panel