completed the navbar
completed the navbar. added a dropdown menu for smaller screen sizes (mobile), gave it the OG UL. Also gave it some styling, gave it gradient and brought it down 20px and with other alignment properties. edited the font and text size for the dropdown, while assuring the options closed the menu when clicked and vice versa. Setting the link clicked as the active page.